List Reorder and Icon Fix
Put all missing icon headsets under the ones with new icons, and fixed the E3's icon having no icon due to a slight error when copy-pasting my code in. Forgot the "i" at the beginning of "invert"

I changed the priorities of all the headsets I reordered, so if that breaks it I'll change them back to the original priority values while keeping the list reorder.