Added some icons for headsets with missing icons (#262)
* Headset Icons

* Fixed Nvidia Icon

There was like 2 blank pixels on the top and left so I cropped those out

* The nvidia icon broke for some reason

* fixed it

my Chromebook cropping the og SVG corrupted it somehow :sob:

* nvm I got it cropped

i got it but i cant delete the SVG or any files I uploaded help

* Update format.js

* Delete public/assets/nvidia.svg

* List Reorder and Icon Fix

Put all missing icon headsets under the ones with new icons, and fixed the E3's icon having no icon due to a slight error when copy-pasting my code in. Forgot the "i" at the beginning of "invert"

I changed the priorities of all the headsets I reordered, so if that breaks it I'll change them back to the original priority values while keeping the list reorder.

* Update format.js

Don't know if it was from me fully reordering the list but the DPVR and Medion icons broke on player's profile stats where the headset and platform is. Fixed the ordering a little too.

* Matched Medion to other WMR headset names

* HMD Icons .png => .svg

Doing this to see if it'll fix the image stretch the icons have on a player's stats (Ranks, Headset, and Platform) since I saw all the other non-square logos as .svg's

Will upload the Nvidia icon later since I guess .svg's can only be black and white and I don't have access to an HTML editor with live preview at the moment to get the color change just right.

* Updated .pngs (Besides CludXR) to use .svg

* Icon Fixes

Fixed the Medion and VRidge icons cause I didn't realize they had to be black before converting them to .svg's
Made Nvidia icon square so it doesn't stretch to fit on a player's listed main headset

* Updated code for icon changes

* Deleted pngs

* testing the vridge and medion icons

* gonna try and reupload the icons to see if they will appear

* Reuploading the svgs

Praying it works now 🙏

* aight test time

* screw this I'm going back to the .png's 💔

* hopefully it should work now